refactors utilities to individual files
