+++ assets/scripts/components/analytics.js
