+++ assets/scripts/components/component.js
