+++ assets/scripts/components/detect.js
