+++ assets/scripts/components/index.js
@@ -1,2 +1,8 @@
+export * from './analytics';
+export * from './component';
+export * from './component-record';
 export * from './convert-to-inline-svg';
+export * from './detect';
+export * from './lazy-loader';
 export * from './social-share';
+export * from './storage';
