+++ assets/scripts/utilities/find-visible-section.js
@@ -1,0 +1,12 @@
+/**
+ * @type public
+ * @name findVisibleSection
+ * @description
+ *
+ * @param {Element} element
+ * @return {Element}
+ *
+ **/
+export function findVisibleSection(element) {
+  return element.is(':hidden') || element.hasClass('gap') ? findVisibleSection(element.next()) : element;
+}
