+++ assets/scripts/utilities/index.js
@@ -1,0 +1,14 @@
+export * from './screen-sizes';
+export * from './tab-index';
+export * from './get-page-type';
+export * from './lock-body';
+export * from './capitalize-first-letter';
+export * from './clear-fragment';
+export * from './shuffle';
+export * from './generate-unique-id';
+export * from './add-script';
+export * from './convert-source-to-inline-svg';
+export * from './ios-fix-double-tap';
+export * from './find-visible-section';
+export * from './move-bootstrap-modals-to-body';
+export * from './ignore-context';
