+++ assets/scripts/utilities/ios-fix-double-tap.js
@@ -1,0 +1,17 @@
+/**
+ * @type public
+ * @name iOSFixDoubleTap
+ * @description
+ *
+ * Fixes double tap issue on iOS devices
+ *
+ * @param {Array} array
+ * @return {Array}
+ *
+ **/
+export function iOSFixDoubleTap() {
+  $('button, a').on('touchstart', (event) => {
+    const $target = $(event.currentTarget);
+    $target.is(':focus') ? $target.blur() : $target.focus();
+  });
+}
