+++ assets/scripts/utilities/shuffle.js
@@ -1,0 +1,30 @@
+/**
+ * @type public
+ * @name shuffle
+ * @description
+ *
+ * Fisher-Yates (aka Knuth) Shuffle.
+ *
+ * @param {Array} array
+ * @return {Array}
+ *
+ **/
+export function shuffle(array) {
+  let currentIndex = array.length;
+  let temporaryValue;
+  let randomIndex;
+
+  // While there remain elements to shuffle...
+  while (0 !== currentIndex) {
+    // Pick a remaining element...
+    randomIndex = Math.floor(Math.random() * currentIndex);
+    currentIndex -= 1;
+
+    // And swap it with the current element.
+    temporaryValue = array[currentIndex];
+    array[currentIndex] = array[randomIndex];
+    array[randomIndex] = temporaryValue;
+  }
+
+  return array;
+}
